requirements.txt
aiohttp==3.7.4.post0
apiclient==1.0.4
APScheduler==3.6.3
asgiref==3.5.2
asttokens==2.0.5
async-timeout==3.0.1
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.4.1
bleach==4.1.0
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
charset-normalizer==2.0.6
click==8.1.3
colorama==0.4.4
comtypes==1.1.11
configparser==5.0.2
crayons==0.4.0
cycler==0.10.0
dataframe-image==0.1.1
decorator==5.1.0
defusedxml==0.7.1
Django==4.1.2
entrypoints==0.3
et-xmlfile==1.1.0
executing==0.8.3
ffmpeg==1.4
fsspec==2022.8.2
google==3.0.0
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.13.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.0
googleapis-common-protos==1.56.4
grpcio==1.49.1
grpcio-tools==1.48.2
gTTS==2.2.4
html-table-parser==0.1.0
httplib2==0.20.4
idna==3.2
ipython==8.3.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.2
JPype1==1.3.0
jsonschema==4.1.0
jupyter-client==7.0.6
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
kiwisolver==1.3.2
konlpy==0.6.0
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
MouseInfo==0.1.3
multidict==5.2.0
multitasking==0.0.10
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
numpy==1.20.2
oauthlib==3.2.2
openpyxl==3.0.9
packaging==21.0
pandas==1.3.5
pandas-datareader==0.10.0
pandocfilters==1.5.0
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.2
prompt-toolkit==3.0.20
protobuf==3.20.3
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
Pygments==2.10.0
PyMsgBox==1.0.9
pyparsing==2.4.7
PyPDF2==1.27.12
PyPDF4==1.27.0
pyperclip==1.8.2
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyQtWebEngine-Qt5==5.15.2
PyRect==0.1.4
pyrsistent==0.18.0
PyScreeze==0.1.28
python-dateutil==2.8.1
python-docx==0.8.11
python-dotenv==0.21.0
python-telegram-bot==13.12
pytube==12.1.0
pytweening==1.0.4
pytz==2021.1
pywin32==300
pywinauto==0.6.8
pyzmq==22.3.0
reportlab==3.6.9
requests==2.26.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.7.1
seaborn==0.11.2
selenium==3.141.0
six==1.15.0
sqlparse==0.4.3
stability-sdk==0.2.4
stack-data==0.2.0
tenacity==8.0.1
testpath==0.5.0
tornado==6.1
traitlets==5.1.0
typing-extensions==3.10.0.2
tzdata==2022.5
tzlocal==2.1
uritemplate==4.1.1
urllib3==1.26.7
wcwidth==0.2.5
webdriver-manager==3.5.0
webencodings==0.5.1
xlrd==2.0.1
yarl==1.7.0
yfinance==0.1.66
